+++ include/wx/osx/window.h
@@ -104,7 +104,7 @@ public:
 
     virtual int GetCharHeight() const;
     virtual int GetCharWidth() const;
-
+    
 public:
     virtual void SetScrollbar( int orient, int pos, int thumbVisible,
                                int range, bool refresh = true );
@@ -299,7 +299,7 @@ public:
 
     bool                IsNativeWindowWrapper() const { return m_isNativeWindowWrapper; }
     
-    float               GetContentScaleFactor() const ;
+    double              GetContentScaleFactor() const ;
     
     // internal response to size events
     virtual void MacOnInternalSize() {}
